# Optimize repeated insertions/deletions from chpl__hashtable